Expanded network validation to the other address.js functions
