+++ routes/v2/address.js
@@ -5,6 +5,7 @@ const router = express.Router()
 const axios = require("axios")
 const RateLimit = require("express-rate-limit")
 const logger = require("./logging.js")
+const routeUtils = require("./route-utils")
 
 // Used for processing error messages before sending them to the user.
 const util = require("util")
@@ -81,7 +82,8 @@ async function details(req, res, next) {
         })
       }
 
-      const networkIsValid = validateNetwork(thisAddress)
+      // Prevent a common user error. Ensure they are using the correct network address.
+      const networkIsValid = routeUtils.validateNetwork(thisAddress)
       if (!networkIsValid) {
         res.status(400)
         return res.json({
@@ -150,6 +152,15 @@ async function utxo(req, res, next) {
         })
       }
 
+      // Prevent a common user error. Ensure they are using the correct network address.
+      const networkIsValid = routeUtils.validateNetwork(thisAddress)
+      if (!networkIsValid) {
+        res.status(400)
+        return res.json({
+          error: `Invalid network. Trying to use a testnet address on mainnet, or vice versa.`
+        })
+      }
+
       const path = `${process.env.BITCOINCOM_BASEURL}addr/${legacyAddr}/utxo`
 
       // Query the Insight server.
@@ -203,6 +214,15 @@ async function unconfirmed(req, res, next) {
         res.status(400)
         return res.json({
           error: `Invalid BCH address. Double check your address is valid: ${thisAddress}`
+        })
+      }
+
+      // Prevent a common user error. Ensure they are using the correct network address.
+      const networkIsValid = routeUtils.validateNetwork(thisAddress)
+      if (!networkIsValid) {
+        res.status(400)
+        return res.json({
+          error: `Invalid network. Trying to use a testnet address on mainnet, or vice versa.`
         })
       }
 
@@ -268,6 +288,15 @@ async function transactions(req, res, next) {
         })
       }
 
+      // Prevent a common user error. Ensure they are using the correct network address.
+      const networkIsValid = routeUtils.validateNetwork(thisAddress)
+      if (!networkIsValid) {
+        res.status(400)
+        return res.json({
+          error: `Invalid network. Trying to use a testnet address on mainnet, or vice versa.`
+        })
+      }
+
       const path = `${
         process.env.BITCOINCOM_BASEURL
       }txs/?address=${thisAddress}`
@@ -295,37 +324,6 @@ async function transactions(req, res, next) {
     if (error.response && error.response.data && error.response.data.error)
       return res.json({ error: error.response.data.error })
     return res.json({ error: util.inspect(error) })
-  }
-}
-
-// Returns true if user-provided cash address matches the correct network,
-// mainnet or testnet. If NETWORK env var is not defined, it returns false.
-function validateNetwork(addr) {
-  try {
-    const network = process.env.NETWORK
-
-    // Return false if NETWORK is not defined.
-    if (!network || network === "") {
-      console.log(`Warning: NETWORK environment variable is not defined!`)
-      return false
-    }
-
-    // Convert the user-provided address to a cashaddress, for easy detection
-    // of the intended network.
-    const cashAddr = BITBOX.Address.toCashAddress(addr)
-
-    // Return true if the network and address both match testnet
-    const addrIsTest = cashAddr.indexOf(`bchtest`) > -1
-    if (network === "testnet" && addrIsTest) return true
-
-    // Return true if the network and address both match mainnet
-    const addrIsMain = cashAddr.indexOf(`bitcoincash`) > -1
-    if (network === "mainnet" && addrIsMain) return true
-
-    return false
-  } catch (err) {
-    logger.error(`Error in validateNetwork()`)
-    return false
   }
 }
